+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased]
+----------------------
+##### Fixed
+- CSS build issue with Grid Theme.
+
 [4.39.2] - 2017-09-26
 ----------------------
 ##### Fixed